Add status field to user mentions. Change behavior so only old mentions of the same status are destroyed on regeneration

+++ lib/code_corps/scenario/generate_user_mentions_for_comment.rb
@@ -4,39 +4,67 @@ module CodeCorps
       def initialize(comment)
         @comment = comment
         @post = comment.post
+        @publishing = comment.publishing
       end
+
+      attr_reader :comment, :post, :publishing
+      alias_method :publishing?, :publishing
 
       def call
-        result = []
-        mentions = []
-
-        @comment.body_preview.scan(/\B@((?:(?:(?:[^-\W]-?))*)(?:[^\/\W]\/?)?(?:(?:(?:[^-\W]-?))*)\w+)/) do |temp|
-          username = temp.first
-          start_index = Regexp.last_match.offset(0).first
-          end_index = start_index + username.length
-          result << [ username, [start_index, end_index] ]
-        end
-
-        users = User.where(username: result.map(&:first))
-
-        result.each do |r|
-          users.each do |u|
-            if r.first == u.username
-              mentions << [u, r.last.first, r.last.last]
-            end
-          end
-        end
-
         ActiveRecord::Base.transaction do
-          existing_mentions = CommentUserMention.where(comment: @comment)
-          if existing_mentions.present?
-            existing_mentions.destroy_all
-          end
+          destroy_existing_mentions
           mentions.each do |m|
-            CommentUserMention.create!(comment: @comment, post: @post, user: m[0], start_index: m[1], end_index: m[2], username: m[0].username)
+            CommentUserMention.create!(
+              comment: comment, post: post, user: m[0], status: status,
+              start_index: m[1], end_index: m[2], username: m[0].username
+            )
           end
         end
       end
+
+      private
+
+        def status
+          publishing ? :published : :preview
+        end
+
+        def destroy_existing_mentions
+          existing_mentions = comment.comment_user_mentions.published if publishing?
+          existing_mentions = comment.comment_user_mentions.preview unless publishing?
+          existing_mentions.destroy_all if existing_mentions.present?
+        end
+
+        def regex_matches
+          regex = %r{\B@((?:(?:(?:[^-\W]-?))*)(?:[^\/\W]\/?)?(?:(?:(?:[^-\W]-?))*)\w+)}
+
+          result = []
+
+          content = publishing ? comment.body : comment.body_preview
+
+          content.scan(regex) do |temp|
+            username = temp.first
+            start_index = Regexp.last_match.offset(0).first
+            end_index = start_index + username.length
+            result << [username, [start_index, end_index]]
+          end
+
+          result
+        end
+
+        def mentions
+          matches = regex_matches
+          users = User.where(username: matches.map(&:first))
+
+          result = []
+
+          matches.each do |r|
+            users.each do |u|
+              result << [u, r.last.first, r.last.last] if r.first == u.username
+            end
+          end
+
+          result
+        end
     end
   end
 end
